devnest-gateway/requirements-test.txt
# DevNest gateway CI / local test dependencies (not runtime).
pytest>=8.0
PyYAML>=6.0
